+++ CHANGES.rst
@@ -7,6 +7,10 @@ Changelog of threedi-modelchecker
 
 - Try to add `start_datetime` and `duration` to a `GlobalSettingOption`
 
+- Pass temporary database copy file in a different context so it is opened
+  one time, previously it was opened twice which results in errors on Windows.
+  
+
 0.25.2 (2022-01-26)
 -------------------